requirements.txt
asgiref==3.8.1
boto3==1.37.28
botocore==1.37.28
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-allauth==65.7.0
django-axes==7.0.2
django-cors-headers==4.7.0
django-crontab==0.7.1
django-ratelimit==4.1.0
django-rest-passwordreset==1.5.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
pip==25.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
